Move createTx to Tx API
